Changed: Merged structures, exceptions and constants from ixxat-fd to ixxat.
