--- can/interfaces/ixxat_fd/__init__.py
@@ -1,7 +1,0 @@
-"""
-Ctypes wrapper module for IXXAT Virtual CAN Interface V3 on win32 systems
-
-Copyright (C) 2016 Giuseppe Corbelli <giuseppe.corbelli@weightpack.com>
-"""
-
-from can.interfaces.ixxat_fd.canlib import IXXATBus, get_ixxat_hwids
